--- a/pkg/streamingpromql/operators/binops/binary_operation.go
+++ b/pkg/streamingpromql/operators/binops/binary_operation.go
@@ -3,15 +3,19 @@
 package binops
 
 import (
+	"fmt"
 	"slices"
 
 	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/model/labels"
+	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/promql/parser"
 	"github.com/prometheus/prometheus/promql/parser/posrange"
 	"github.com/prometheus/prometheus/util/annotations"
 
+	"github.com/grafana/mimir/pkg/streamingpromql/compat"
 	"github.com/grafana/mimir/pkg/streamingpromql/limiting"
+	"github.com/grafana/mimir/pkg/streamingpromql/operators/functions"
 	"github.com/grafana/mimir/pkg/streamingpromql/types"
 )
 
@@ -43,6 +47,35 @@ func vectorMatchingGroupKeyFunc(vectorMatching parser.VectorMatching) func(label
 
 	return func(l labels.Labels) []byte {
 		return l.BytesWithoutLabels(buf, lbls...)
+	}
+}
+
+// vectorMatchingGroupLabelsFunc returns a function that computes the labels of the output group a series belongs to.
+func groupLabelsFunc(vectorMatching parser.VectorMatching, returnBool bool) func(labels.Labels) labels.Labels {
+	lb := labels.NewBuilder(labels.EmptyLabels())
+
+	if vectorMatching.On {
+		return func(l labels.Labels) labels.Labels {
+			lb.Reset(l)
+			lb.Keep(vectorMatching.MatchingLabels...)
+			return lb.Labels()
+		}
+	}
+
+	if returnBool {
+		// If this is a comparison operator, we want to retain the metric name, as the comparison acts like a filter.
+		return func(l labels.Labels) labels.Labels {
+			lb.Reset(l)
+			lb.Del(vectorMatching.MatchingLabels...)
+			return lb.Labels()
+		}
+	}
+
+	return func(l labels.Labels) labels.Labels {
+		lb.Reset(l)
+		lb.Del(labels.MetricName)
+		lb.Del(vectorMatching.MatchingLabels...)
+		return lb.Labels()
 	}
 }
 
@@ -114,4 +147,199 @@ func sampleTypeDescription(h *histogram.FloatHistogram) string {
 	}
 
 	return "histogram"
+}
+
+type vectorVectorBinaryOperationEvaluator struct {
+	op                       parser.ItemType
+	opFunc                   binaryOperationFunc
+	leftIterator             types.InstantVectorSeriesDataIterator
+	rightIterator            types.InstantVectorSeriesDataIterator
+	memoryConsumptionTracker *limiting.MemoryConsumptionTracker
+	annotations              *annotations.Annotations
+	expressionPosition       posrange.PositionRange
+}
+
+func newVectorVectorBinaryOperationEvaluator(
+	op parser.ItemType,
+	returnBool bool,
+	memoryConsumptionTracker *limiting.MemoryConsumptionTracker,
+	annotations *annotations.Annotations,
+	expressionPosition posrange.PositionRange,
+) (vectorVectorBinaryOperationEvaluator, error) {
+	e := vectorVectorBinaryOperationEvaluator{
+		op:                       op,
+		opFunc:                   nil,
+		memoryConsumptionTracker: memoryConsumptionTracker,
+		annotations:              annotations,
+		expressionPosition:       expressionPosition,
+	}
+
+	if returnBool {
+		e.opFunc = boolComparisonOperationFuncs[op]
+	} else {
+		e.opFunc = arithmeticAndComparisonOperationFuncs[op]
+	}
+
+	if e.opFunc == nil {
+		return vectorVectorBinaryOperationEvaluator{}, compat.NewNotSupportedError(fmt.Sprintf("binary expression with '%s'", op))
+	}
+
+	return e, nil
+
+}
+
+func (e *vectorVectorBinaryOperationEvaluator) computeResult(left types.InstantVectorSeriesData, right types.InstantVectorSeriesData, takeOwnershipOfLeft bool, takeOwnershipOfRight bool) (types.InstantVectorSeriesData, error) {
+	var fPoints []promql.FPoint
+	var hPoints []promql.HPoint
+
+	// For one-to-one matching for arithmetic operators, we'll never produce more points than the smaller input side.
+	// Because floats and histograms can be multiplied together, we use the sum of both the float and histogram points.
+	// We also don't know if the output will be exclusively floats or histograms, so we'll use the same size slice for both.
+	// We only assign the slices once we see the associated point type so it shouldn't be common that we allocate both.
+	canReturnLeftFPointSlice, canReturnLeftHPointSlice, canReturnRightFPointSlice, canReturnRightHPointSlice := takeOwnershipOfLeft, takeOwnershipOfLeft, takeOwnershipOfRight, takeOwnershipOfRight
+	leftPoints := len(left.Floats) + len(left.Histograms)
+	rightPoints := len(right.Floats) + len(right.Histograms)
+	maxPoints := max(leftPoints, rightPoints)
+
+	// We cannot re-use any slices when the series contain a mix of floats and histograms.
+	// Consider the following, where f is a float at a particular step, and h is a histogram.
+	// load 5m
+	//   series1 f f f h h
+	//   series2 h h f f h
+	// eval range from 0 to 25m step 5m series1 * series2
+	//   {}      h h f h f
+	// We can fit the resulting 3 histograms into series2 existing slice. However, the second
+	// last step (index 3) produces a histogram which would be stored over the existing histogram
+	// at the end of series2 (also index 3).
+	// It should be pretty uncommon that metric contains both histograms and floats, so we will
+	// accept the cost of a new slice.
+	mixedPoints := (len(left.Floats) > 0 && len(left.Histograms) > 0) || (len(right.Floats) > 0 && len(right.Histograms) > 0)
+
+	prepareFSlice := func() error {
+		if !mixedPoints && maxPoints <= cap(left.Floats) && cap(left.Floats) < cap(right.Floats) && takeOwnershipOfLeft {
+			// Can fit output in left side, the left side is smaller than the right, and we're allowed to modify it
+			canReturnLeftFPointSlice = false
+			fPoints = left.Floats[:0]
+			return nil
+		}
+		if !mixedPoints && maxPoints <= cap(right.Floats) && takeOwnershipOfRight {
+			// Can otherwise fit in the right side and we're allowed to modify it
+			canReturnRightFPointSlice = false
+			fPoints = right.Floats[:0]
+			return nil
+		}
+		// Either we have mixed points or we can't fit in either left or right side, so create a new slice
+		var err error
+		if fPoints, err = types.FPointSlicePool.Get(maxPoints, e.memoryConsumptionTracker); err != nil {
+			return err
+		}
+		return nil
+	}
+
+	prepareHSlice := func() error {
+		if !mixedPoints && maxPoints <= cap(left.Histograms) && cap(left.Histograms) < cap(right.Histograms) && takeOwnershipOfLeft {
+			// Can fit output in left side, the left side is smaller than the right, and we're allowed to modify it
+			canReturnLeftHPointSlice = false
+			hPoints = left.Histograms[:0]
+			return nil
+		}
+		if !mixedPoints && maxPoints <= cap(right.Histograms) && takeOwnershipOfRight {
+			// Can otherwise fit in the right side and we're allowed to modify it
+			canReturnRightHPointSlice = false
+			hPoints = right.Histograms[:0]
+			return nil
+		}
+		// Either we have mixed points or we can't fit in either left or right side, so create a new slice
+		var err error
+		if hPoints, err = types.HPointSlicePool.Get(maxPoints, e.memoryConsumptionTracker); err != nil {
+			return err
+		}
+		return nil
+	}
+
+	e.leftIterator.Reset(left)
+	e.rightIterator.Reset(right)
+
+	// Get first sample from left and right
+	lT, lF, lH, lOk := e.leftIterator.Next()
+	rT, rF, rH, rOk := e.rightIterator.Next()
+	// Continue iterating until we exhaust either the LHS or RHS
+	// denoted by lOk or rOk being false.
+	for lOk && rOk {
+		if lT == rT {
+			// We have samples on both sides at this timestep.
+			resultFloat, resultHist, keep, valid, err := e.opFunc(lF, rF, lH, rH)
+
+			if err != nil {
+				err = functions.NativeHistogramErrorToAnnotation(err, e.emitAnnotation)
+				if err != nil {
+					return types.InstantVectorSeriesData{}, err
+				}
+
+				// Else: error was converted to an annotation, continue without emitting a sample here.
+				keep = false
+			}
+
+			if !valid {
+				emitIncompatibleTypesAnnotation(e.annotations, e.op, lH, rH, e.expressionPosition)
+			}
+
+			if keep {
+				if resultHist != nil {
+					if hPoints == nil {
+						if err = prepareHSlice(); err != nil {
+							return types.InstantVectorSeriesData{}, err
+						}
+					}
+					hPoints = append(hPoints, promql.HPoint{
+						H: resultHist,
+						T: lT,
+					})
+				} else {
+					if fPoints == nil {
+						if err = prepareFSlice(); err != nil {
+							return types.InstantVectorSeriesData{}, err
+						}
+					}
+					fPoints = append(fPoints, promql.FPoint{
+						F: resultFloat,
+						T: lT,
+					})
+				}
+			}
+		}
+
+		// Advance the iterator with the lower timestamp, or both if equal
+		if lT == rT {
+			lT, lF, lH, lOk = e.leftIterator.Next()
+			rT, rF, rH, rOk = e.rightIterator.Next()
+		} else if lT < rT {
+			lT, lF, lH, lOk = e.leftIterator.Next()
+		} else {
+			rT, rF, rH, rOk = e.rightIterator.Next()
+		}
+	}
+
+	// Cleanup the unused slices.
+	if canReturnLeftFPointSlice {
+		types.FPointSlicePool.Put(left.Floats, e.memoryConsumptionTracker)
+	}
+	if canReturnLeftHPointSlice {
+		types.HPointSlicePool.Put(left.Histograms, e.memoryConsumptionTracker)
+	}
+	if canReturnRightFPointSlice {
+		types.FPointSlicePool.Put(right.Floats, e.memoryConsumptionTracker)
+	}
+	if canReturnRightHPointSlice {
+		types.HPointSlicePool.Put(right.Histograms, e.memoryConsumptionTracker)
+	}
+
+	return types.InstantVectorSeriesData{
+		Floats:     fPoints,
+		Histograms: hPoints,
+	}, nil
+}
+
+func (e *vectorVectorBinaryOperationEvaluator) emitAnnotation(generator types.AnnotationGenerator) {
+	e.annotations.Add(generator("", e.expressionPosition))
 }