… cmd/mimir/config-descriptor.json
… docs/sources/mimir/configure/configuration-parameters/index.md
… pkg/streamingpromql/benchmarks/benchmarks.go
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/operators/binops/grouped_vector_vector_binary_operation_test.go
… pkg/streamingpromql/operators/binops/one_to_one_vector_vector_binary_operation.go
… pkg/streamingpromql/operators/binops/one_to_one_vector_vector_binary_operation_test.go
… pkg/streamingpromql/operators/binops/vector_vector_binary_operation.go
… pkg/streamingpromql/query.go
… pkg/streamingpromql/testdata/ours/binary_operators.test
… pkg/streamingpromql/testdata/upstream/collision.test
… pkg/streamingpromql/testdata/upstream/operators.test
… pkg/streamingpromql/types/limiting_pool.go