#### What this PR does

This PR adds support for `group_left` and `group_right` (aka many-to-one and one-to-many matching) in binary operations to MQE.

In our benchmarks, latency is up to 80% lower and peak memory utilisation is up to 30% lower with MQE compared to Prometheus' engine:

```
                                                                         │  Prometheus  │               Mimir                │
                                                                         │    sec/op    │   sec/op     vs base               │
Query/h_1_*_on(l)_group_left()_a_1,_instant_query-10                        321.8µ ± 2%   340.3µ ± 9%        ~ (p=0.065 n=6)
Query/h_1_*_on(l)_group_left()_a_1,_range_query_with_100_steps-10           452.3µ ± 1%   358.5µ ± 2%  -20.74% (p=0.002 n=6)
Query/h_1_*_on(l)_group_left()_a_1,_range_query_with_1000_steps-10         1651.5µ ± 0%   756.0µ ± 3%  -54.23% (p=0.002 n=6)
Query/h_100_*_on(l)_group_left()_a_100,_instant_query-10                    4.694m ± 1%   4.659m ± 2%        ~ (p=0.310 n=6)
Query/h_100_*_on(l)_group_left()_a_100,_range_query_with_100_steps-10      22.581m ± 1%   8.320m ± 1%  -63.16% (p=0.002 n=6)
Query/h_100_*_on(l)_group_left()_a_100,_range_query_with_1000_steps-10     185.37m ± 3%   39.77m ± 0%  -78.55% (p=0.002 n=6)
Query/h_2000_*_on(l)_group_left()_a_2000,_instant_query-10                  84.71m ± 1%   85.24m ± 2%        ~ (p=0.132 n=6)
Query/h_2000_*_on(l)_group_left()_a_2000,_range_query_with_100_steps-10     468.3m ± 1%   155.0m ± 4%  -66.90% (p=0.002 n=6)
Query/h_2000_*_on(l)_group_left()_a_2000,_range_query_with_1000_steps-10   4006.0m ± 2%   761.7m ± 1%  -80.99% (p=0.002 n=6)
geomean                                                                     20.87m        10.41m       -50.14%

                                                                         │   Prometheus   │                Mimir                │
                                                                         │       B        │      B        vs base               │
Query/h_1_*_on(l)_group_left()_a_1,_instant_query-10                        73.48Mi ±  1%   73.76Mi ± 1%        ~ (p=0.699 n=6)
Query/h_1_*_on(l)_group_left()_a_1,_range_query_with_100_steps-10           72.05Mi ±  1%   73.21Mi ± 1%   +1.62% (p=0.002 n=6)
Query/h_1_*_on(l)_group_left()_a_1,_range_query_with_1000_steps-10          70.13Mi ±  1%   70.78Mi ± 1%        ~ (p=0.087 n=6)
Query/h_100_*_on(l)_group_left()_a_100,_instant_query-10                    69.96Mi ±  1%   70.16Mi ± 1%   +0.29% (p=0.015 n=6)
Query/h_100_*_on(l)_group_left()_a_100,_range_query_with_100_steps-10       73.55Mi ±  1%   73.12Mi ± 1%        ~ (p=0.262 n=6)
Query/h_100_*_on(l)_group_left()_a_100,_range_query_with_1000_steps-10     107.05Mi ±  5%   90.28Mi ± 3%  -15.66% (p=0.002 n=6)
Query/h_2000_*_on(l)_group_left()_a_2000,_instant_query-10                  79.68Mi ±  3%   79.24Mi ± 4%        ~ (p=0.485 n=6)
Query/h_2000_*_on(l)_group_left()_a_2000,_range_query_with_100_steps-10     172.9Mi ±  6%   131.9Mi ± 2%  -23.73% (p=0.002 n=6)
Query/h_2000_*_on(l)_group_left()_a_2000,_range_query_with_1000_steps-10    651.3Mi ± 13%   438.1Mi ± 0%  -32.74% (p=0.002 n=6)
geomean                                                                     107.0Mi         97.69Mi        -8.68%
```

I am not adding a changelog entry given it is covered by the generic MQE changelog entry that links to https://github.com/grafana/mimir/issues/10067.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/issues/10067

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.