@jhesketh 2024-12-05
Forgot to add that I think we should see group left/right in the mixed metrics test gauntlet