@charleskorn 2024-12-10
> Forgot to add that I think we should see group left/right in the mixed metrics test gauntlet

Added in 7214a8e.